requirements.txt
cytoolz==0.11.0
fvcore==0.1.1.post20200704
numpy==1.18.1
pytorch-ignite==0.5.0.dev20200823
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.0
tensorboard==2.1.1
torchdiffeq==0.2.0
tqdm